+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 
 # 0.12.2
 
+- Peers now join the pool progressively
 - Save password and grid name to keep conf on node restart
+- Set GRIDFILE env